+++ test/QueryTest.php
@@ -22,6 +22,27 @@ class QueryTest extends PHPUnit_Framework_TestCase
         $this->query = new Query('kingkong=toto');
     }
 
+    /**
+     * @param $str
+     * @expectedException InvalidArgumentException
+     * @dataProvider failedConstructor
+     */
+    public function testFailedConstructor($str)
+    {
+        new Query($str);
+    }
+
+    public function failedConstructor()
+    {
+        return [
+            'bool'      => [true],
+            'Std Class' => [(object) 'foo'],
+            'null'      => [null],
+            'float'     => [1.2],
+            'array'      =>[['foo']],
+        ];
+    }
+
     public function testIterator()
     {
         $this->assertSame(['kingkong' => 'toto'], iterator_to_array($this->query, true));